+++ Krowi_AchievementFilter/GUI/WorldMapButton.lua
@@ -9,19 +9,9 @@ local worldMapButton = gui.WorldMapButton;
 addon.WorldMapButtons = LibStub("KrowiWorldMapButtons-1.0"); -- Global world map buttons object
 
 function worldMapButton.Load()
-    addon.WorldMapButtons:Add("TemplateName (need to write)", "BUTTON");
+    addon.WorldMapButtons:Add("KrowiAF_WorldMapAchievementButtonTemplate", "BUTTON");
 end
 
--- function addon.FixHandyNotesCompatibility()
---     for _, f in next, WorldMapFrame.overlayFrames do
---         if f.OnLoad == HandyNotes_ShadowlandsWorldMapOptionsButtonMixin.OnLoad then
---             print("found");
---         end
---         if f.OnLoad == HandyNotes_VisionsOfNZothWorldMapOptionsButtonMixin.OnLoad then
---             print("found");
---         end
---     end
--- end
 
 WorldMapAchievementButtonMixin = {};